modified:   src/main/frontend/src/App.js 	modified:   src/main/frontend/src/PlaceRecommend/PlaceRecommend.js 	renamed:    src/main/frontend/src/AboutUs.js -> src/main/frontend/src/aboutUs/AboutUs.js 	renamed:    src/main/frontend/src/Community.js -> src/main/frontend/src/community/Community.js 	modified:   src/main/frontend/src/home/Home.js 	modified:   src/main/frontend/src/index.js 	renamed:    src/main/frontend/src/Bumper.js -> src/main/frontend/src/layout/Bumper.js 	renamed:    src/main/frontend/src/NavVillage.js -> src/main/frontend/src/layout/NavVillage.js 	renamed:    src/main/frontend/src/sideBar/SideBar.js -> src/main/frontend/src/layout/SideBar.js 	renamed:    src/main/frontend/src/footer.js -> src/main/frontend/src/layout/footer.js 	renamed:    src/main/frontend/src/sideBar/sideBarStyle.css -> src/main/frontend/src/layout/sideBarStyle.css 	modified:   src/main/frontend/src/membership/JoinMembership.js 	renamed:    src/main/frontend/src/Mypage.js -> src/main/frontend/src/membership/Mypage.js 	modified:   src/main/frontend/src/test.js 	modified:   src/main/frontend/src/App.js 	modified:   src/main/frontend/src/test.js
