+++ src/main/frontend/src/layout/footer.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import "./style.css"
+import "../style.css"
 import { FaXTwitter, FaFacebookF, FaInstagram, FaGithub } from "react-icons/fa6";
 
 const FooterContainer = styled.div`
